保存文章和获取文章

+++ src/main/java/top/joylife/lemon/dao/impl/ArticleTagDaoImpl.java
@@ -2,6 +2,7 @@ package top.joylife.lemon.dao.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+import org.springframework.util.CollectionUtils;
 import tk.mybatis.mapper.entity.Example;
 import top.joylife.lemon.dao.ArticleTagDao;
 import top.joylife.lemon.dao.entity.ArticleTag;
@@ -36,6 +37,9 @@ public class ArticleTagDaoImpl implements ArticleTagDao {
      */
     @Override
     public void batchDelete(List<Integer> ids) {
+        if(!CollectionUtils.isEmpty(ids)){
+            return;
+        }
         Example example = new Example(ArticleTag.class);
         example.createCriteria()
                 .andIn("id",ids);
@@ -49,9 +53,26 @@ public class ArticleTagDaoImpl implements ArticleTagDao {
      */
     @Override
     public void deleteByArticleId(Integer articleId) {
+        if(articleId==null){
+            return;
+        }
         Example example = new Example(ArticleTag.class);
         example.createCriteria()
                 .andEqualTo("articleId",articleId);
         articleTagMapper.deleteByExample(example);
+    }
+
+    /**
+     * 根据文章id查询
+     *
+     * @param articleId
+     * @return
+     */
+    @Override
+    public List<ArticleTag> listByArticleId(Integer articleId) {
+        Example example = new Example(ArticleTag.class);
+        example.createCriteria()
+                .andEqualTo("articleId",articleId);
+        return articleTagMapper.selectByExample(example);
     }
 }
